Properties added: active, emptyView, transparentBackground, minAlpha
Allow to disable fading effect if needed.
Allow to specify that the view is empty for measurement issues.
Allow to set a minimum alpha value greater that zero.
Allow to set a transparent background for the content view.

Added ability for the content view to fill the screen if smaller.

Min SDK version incremented.